fix: Increase PageSize of ListPolicies Paginator (#2033)
SAM runs within a Lambda function and loads IAM Managed Policies once per
Lambda. Previous to this, SAM would call IAM 9 times which could cause
throttling by IAM. With this change, we update the MaxItems from the
default of 100 to the max (1000). In local testing, this has shown
a 0.6 second reduction in the latency in calling IAM.

Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>